--- rework_pysatl_mpest/Initializers/experiments/em_pipeline_results/em_pipeline_comparison_aggregated_results.csv
@@ -1,19 +1,0 @@
-method,n_samples,kl_divergence_mean,kl_divergence_std,final_log_likelihood_mean,final_log_likelihood_std,total_iterations_mean,total_iterations_std,execution_time_mean,execution_time_std,total_time_mean,total_time_std
-EM,100,0.09160780005184557,0.023991342354113248,-150.57564713363632,15.623792248747439,500.0,0.0,0.07439932823181153,0.04513197669742934,0.07439932823181153,0.04513197669742934
-Kmeans,100,0.09391592052246671,0.03016253121494317,-148.39589714331217,13.61090875687515,500.0,0.0,0.05332303047180176,0.0006788732815116693,0.08435087203979492,0.059741662049384005
-Cmeans,100,0.091321956217046,0.0243297321211135,-149.4521232077257,16.365154444883593,500.0,0.0,0.05362858772277832,0.0014417407115280515,0.055225038528442384,0.000875583404213466
-EM,200,0.04669543339864479,0.008530641802334088,-307.76073071654065,11.860049221786012,500.0,0.0,0.05788364410400391,0.0008699490577554425,0.05788364410400391,0.0008699490577554425
-Kmeans,200,0.0484009194712121,0.005411254946931905,-308.03887092694487,12.275688890028372,500.0,0.0,0.05822939872741699,0.0005818098596563509,0.06225333213806152,0.0005448337331994376
-Cmeans,200,0.04674884565026455,0.008423647024040767,-308.62539147751914,13.205401072168529,500.0,0.0,0.0576270580291748,0.0004293940320084507,0.05946846008300781,0.0004095719761938587
-EM,300,0.04297593225028246,0.01010173501257558,-461.98326956217954,14.45042446071576,500.0,0.0,0.06102108955383301,0.0003914736146575062,0.06102108955383301,0.0003914736146575062
-Kmeans,300,0.04166303459853103,0.011441740501171634,-462.4882853953779,14.719466005680056,500.0,0.0,0.060765838623046874,0.0003580661551656588,0.06654558181762696,0.0008333114688787562
-Cmeans,300,0.041663045453458156,0.011441729365659731,-462.48828513303505,14.719465596061704,500.0,0.0,0.06087174415588379,0.0008048636326883679,0.0627263069152832,0.0007913960313463473
-EM,400,0.0407004976290588,0.0024349618289170824,-613.078627450967,27.527514925354,500.0,0.0,0.06478462219238282,0.0005154881119528596,0.06478462219238282,0.0005154881119528596
-Kmeans,400,0.040703076606934036,0.0024352819682761364,-613.2382858195135,27.67625613581049,500.0,0.0,0.06526532173156738,0.0009289178354163659,0.07257981300354004,0.0010993468453783736
-Cmeans,400,0.040700568057722444,0.0024348376155910205,-613.2356454643717,27.653144065082234,500.0,0.0,0.06351842880249023,0.0008756674399253818,0.06608614921569825,0.000321947661155479
-EM,500,0.03884915485052891,0.005516613505029809,-770.952695909997,20.78781920783158,500.0,0.0,0.07280268669128417,0.010042904843304369,0.07280268669128417,0.010042904843304369
-Kmeans,500,0.038957570881678893,0.005366108092232148,-768.1104537012868,23.86292789951412,500.0,0.0,0.07647118568420411,0.005122934605086662,0.08646759986877442,0.008205873653854327
-Cmeans,500,0.03884915809788629,0.0055166092532635165,-770.952695907966,20.787819209496316,500.0,0.0,0.07101516723632813,0.005193410003428706,0.07358055114746094,0.004848633695933944
-EM,1000,0.03490409285000489,0.0029556473265054534,-1562.5377808059875,27.838976015127667,500.0,0.0,0.09068093299865723,0.005764886912887428,0.09068093299865723,0.005764886912887428
-Kmeans,1000,0.03530256176580529,0.0024332653106895956,-1562.4140841613328,27.583656290447596,500.0,0.0,0.08625860214233398,0.0015267030475890236,0.10135960578918457,0.006166425128769406
-Cmeans,1000,0.0349041216413348,0.002955603174419954,-1562.537774608114,27.838963212214196,500.0,0.0,0.0820885181427002,0.0023887052342074855,0.08677759170532226,0.002369859158535294
