--- rework_pysatl_mpest/Initializers/experiments/results/best_methods.csv
@@ -1,21 +1,0 @@
-n_samples,metric,method,silhouette,db,kl_divergence,time
-200,best_db,cmeans_fast,0.6296 ± 0.0107,0.4865 ± 0.0056,0.2320 ± 0.1099,0.0013 ± 0.0002
-200,best_db,kmeans_fast,0.6274 ± 0.0231,0.4978 ± 0.0054,0.3138 ± 0.0400,0.0031 ± 0.0005
-200,best_kl,cmeans_accurate,0.6325 ± 0.0098,0.5008 ± 0.0238,0.1827 ± 0.1029,0.0021 ± 0.0007
-200,best_silhouette,cmeans_fast,0.6331 ± 0.0112,0.4909 ± 0.0098,0.2010 ± 0.1091,0.0013 ± 0.0002
-200,best_silhouette,kmeans_fast,0.6437,0.5016,0.2855,0.0035
-500,best_db,cmeans_fast,0.6174,0.5046,0.2656,0.0018
-500,best_db,kmeans_accurate,0.5969,0.5607,0.1208,0.0102
-500,best_db,kmeans_fast,0.6192 ± 0.0078,0.5208 ± 0.0320,0.2398 ± 0.0942,0.0072 ± 0.0015
-500,best_kl,cmeans_accurate,0.6174,0.5046,0.2656,0.0049
-500,best_kl,cmeans_fast,0.5968 ± 0.0047,0.5595 ± 0.0000,0.2707 ± 0.0050,0.0037 ± 0.0018
-500,best_kl,kmeans_fast,0.6072 ± 0.0149,0.5364 ± 0.0367,0.1258 ± 0.0074,0.0065 ± 0.0018
-500,best_silhouette,cmeans_fast,0.6174,0.5046,0.2656,0.0018
-500,best_silhouette,kmeans_accurate,0.5969,0.5607,0.1208,0.0102
-500,best_silhouette,kmeans_fast,0.6192 ± 0.0078,0.5208 ± 0.0320,0.2398 ± 0.0942,0.0072 ± 0.0015
-1000,best_db,cmeans_fast,0.6078,0.5411,0.2639,0.0042
-1000,best_db,kmeans_fast,0.6338 ± 0.0069,0.5188 ± 0.0152,0.1527 ± 0.0821,0.0141 ± 0.0016
-1000,best_kl,cmeans_accurate,0.6139 ± 0.0036,0.5407 ± 0.0389,0.0976 ± 0.0065,0.0162 ± 0.0012
-1000,best_kl,cmeans_fast,0.6163 ± 0.0119,0.5299 ± 0.0158,0.1792 ± 0.1198,0.0047 ± 0.0008
-1000,best_kl,kmeans_accurate,0.6352,0.5210,0.1105,0.0261
-1000,best_silhouette,kmeans_fast,0.6358 ± 0.0073,0.5251 ± 0.0193,0.1763 ± 0.0886,0.0142 ± 0.0014
