--- rework_pysatl_mpest/Initializers/experiments/results/best_methods_detailed.csv
@@ -1,46 +1,0 @@
-n_samples,metric,method,silhouette,db,kl_divergence,time,experiment
-200,best_silhouette,cmeans_fast,0.625352363295205,0.4806026735365956,0.2931762706528724,0.0011241436004638672,0
-200,best_db,cmeans_fast,0.625352363295205,0.4806026735365956,0.2931762706528724,0.0011241436004638672,0
-200,best_kl,cmeans_accurate,0.625352363295205,0.4806026735365956,0.29315125391898506,0.0016591548919677734,0
-200,best_silhouette,cmeans_fast,0.6217980946703764,0.4916254141455152,0.10506177223876444,0.0012102127075195312,1
-200,best_db,cmeans_fast,0.6217980946703764,0.4916254141455152,0.10506177223876444,0.0012102127075195312,1
-200,best_kl,cmeans_accurate,0.6217980946703764,0.4916254141455152,0.10504711469612396,0.0017910003662109375,1
-200,best_silhouette,cmeans_fast,0.6436008447731869,0.5039107006383748,0.10805094783280486,0.0013990402221679688,2
-200,best_db,kmeans_fast,0.61109513512996,0.49391852556987176,0.34212038163035396,0.0027289390563964844,2
-200,best_kl,cmeans_accurate,0.6436008447731869,0.5039107006383748,0.10803122322217952,0.001878976821899414,2
-200,best_silhouette,cmeans_fast,0.6417727338567366,0.48727936328866933,0.2976679852589723,0.0014941692352294922,3
-200,best_db,cmeans_fast,0.6417727338567366,0.48727936328866933,0.2976679852589723,0.0014941692352294922,3
-200,best_kl,cmeans_accurate,0.6417727338567366,0.48727936328866933,0.29765512868282457,0.0016150474548339844,3
-200,best_silhouette,kmeans_fast,0.6437483034488913,0.5016151959124037,0.2854846766578805,0.0034580230712890625,4
-200,best_db,kmeans_fast,0.6437483034488913,0.5016151959124037,0.2854846766578805,0.0034580230712890625,4
-200,best_kl,cmeans_accurate,0.6301506329544979,0.5404448378739239,0.10984255565599607,0.00333404541015625,4
-500,best_silhouette,kmeans_accurate,0.5968924851730901,0.560748773927632,0.12079626955660881,0.010173797607421875,0
-500,best_db,kmeans_accurate,0.5968924851730901,0.560748773927632,0.12079626955660881,0.010173797607421875,0
-500,best_kl,kmeans_fast,0.5967183114844927,0.5623871162647704,0.12056730205894194,0.005283832550048828,0
-500,best_silhouette,kmeans_fast,0.6177592619314555,0.5104693362108398,0.13106998095686104,0.007774829864501953,1
-500,best_db,kmeans_fast,0.6177592619314555,0.5104693362108398,0.13106998095686104,0.007774829864501953,1
-500,best_kl,kmeans_fast,0.6177592619314555,0.5104693362108398,0.13106998095686104,0.007774829864501953,1
-500,best_silhouette,kmeans_fast,0.6275773152464728,0.49520323759829693,0.29519801281701996,0.008239984512329102,2
-500,best_db,kmeans_fast,0.6275773152464728,0.49520323759829693,0.29519801281701996,0.008239984512329102,2
-500,best_kl,cmeans_fast,0.6000908204261823,0.5594677440316143,0.26717282778004203,0.0025031566619873047,2
-500,best_silhouette,kmeans_fast,0.6122394131556917,0.5567201949786366,0.2932434040173312,0.005471944808959961,3
-500,best_db,kmeans_fast,0.6122394131556917,0.5567201949786366,0.2932434040173312,0.005471944808959961,3
-500,best_kl,cmeans_fast,0.5934198202726243,0.5594629362614074,0.27425981395699917,0.004981040954589844,3
-500,best_silhouette,cmeans_fast,0.6173953229537624,0.5046026904467952,0.26558144127769373,0.0018391609191894531,4
-500,best_db,cmeans_fast,0.6173953229537624,0.5046026904467952,0.26558144127769373,0.0018391609191894531,4
-500,best_kl,cmeans_accurate,0.6173953229537624,0.5046026904467952,0.265581259009776,0.004864215850830078,4
-1000,best_silhouette,kmeans_fast,0.6434209114776361,0.5503062909882207,0.27077929955520297,0.014441967010498047,0
-1000,best_db,cmeans_fast,0.6078313982208688,0.5411187575873747,0.26393411764031366,0.004185199737548828,0
-1000,best_kl,cmeans_fast,0.6078313982208688,0.5411187575873747,0.26393411764031366,0.004185199737548828,0
-1000,best_silhouette,kmeans_fast,0.6284788540191084,0.5408934360902774,0.2758357141161767,0.015885114669799805,1
-1000,best_db,kmeans_fast,0.6284788540191084,0.5408934360902774,0.2758357141161767,0.015885114669799805,1
-1000,best_kl,cmeans_accurate,0.616499796591258,0.5681809911757066,0.09305932007862314,0.017094135284423828,1
-1000,best_silhouette,kmeans_fast,0.6364852254309797,0.5166028036531806,0.11125810018367896,0.014578104019165039,2
-1000,best_db,kmeans_fast,0.6364852254309797,0.5166028036531806,0.11125810018367896,0.014578104019165039,2
-1000,best_kl,kmeans_accurate,0.6351774924786919,0.5210144645699089,0.11047376769250036,0.026097774505615234,2
-1000,best_silhouette,kmeans_fast,0.6423080107444707,0.5089959817068986,0.11316447091153367,0.011963844299316406,3
-1000,best_db,kmeans_fast,0.6423080107444707,0.5089959817068986,0.11316447091153367,0.011963844299316406,3
-1000,best_kl,cmeans_fast,0.6247242401142157,0.5187585317879075,0.09456559804187108,0.005258083343505859,3
-1000,best_silhouette,kmeans_fast,0.6280636810562462,0.5086730321735192,0.11036246175590148,0.014166831970214844,4
-1000,best_db,kmeans_fast,0.6280636810562462,0.5086730321735192,0.11036246175590148,0.014166831970214844,4
-1000,best_kl,cmeans_accurate,0.6113643048244171,0.5131320126384223,0.10220335771400217,0.015368938446044922,4
