--- rework_pysatl_mpest/Initializers/experiments/results/clustering_metrics_results.csv
@@ -1,19 +1,0 @@
-n_samples,method,silhouette,db,kl_divergence,time
-200,cmeans_accurate,0.6325 ± 0.0098,0.5008 ± 0.0238,0.1827 ± 0.1029,0.0021 ± 0.0007
-200,cmeans_fast,0.6325 ± 0.0098,0.5008 ± 0.0238,0.1828 ± 0.1029,0.0013 ± 0.0002
-200,initial,nan,nan,5.8732 ± 5.1260,nan
-200,kmeans_accurate,0.6193 ± 0.0171,0.5132 ± 0.0189,0.3400 ± 0.0736,0.0040 ± 0.0006
-200,kmeans_fast,0.6200 ± 0.0181,0.5084 ± 0.0180,0.3400 ± 0.0736,0.0613 ± 0.1288
-200,true,-0.0644 ± 0.0191,58.0401 ± 106.8217,nan,nan
-500,cmeans_accurate,0.5966 ± 0.0131,0.5478 ± 0.0251,0.2710 ± 0.0046,0.0058 ± 0.0012
-500,cmeans_fast,0.5966 ± 0.0131,0.5477 ± 0.0250,0.2709 ± 0.0046,0.0030 ± 0.0012
-500,initial,nan,nan,9.1348 ± 7.7590,nan
-500,kmeans_accurate,0.6136 ± 0.0111,0.5313 ± 0.0285,0.2267 ± 0.0920,0.0104 ± 0.0015
-500,kmeans_fast,0.6137 ± 0.0112,0.5315 ± 0.0289,0.2267 ± 0.0922,0.0094 ± 0.0062
-500,true,-0.0346 ± 0.0069,56.2738 ± 48.2838,nan,nan
-1000,cmeans_accurate,0.6116 ± 0.0101,0.5395 ± 0.0236,0.1640 ± 0.0924,0.0163 ± 0.0010
-1000,cmeans_fast,0.6116 ± 0.0101,0.5396 ± 0.0236,0.1640 ± 0.0924,0.0048 ± 0.0009
-1000,initial,nan,nan,8.3692 ± 5.1313,nan
-1000,kmeans_accurate,0.6352 ± 0.0070,0.5265 ± 0.0183,0.1762 ± 0.0886,0.0262 ± 0.0012
-1000,kmeans_fast,0.6358 ± 0.0073,0.5251 ± 0.0193,0.1763 ± 0.0886,0.0142 ± 0.0014
-1000,true,-0.0388 ± 0.0090,167.3528 ± 361.3182,nan,nan
