--- rework_pysatl_mpest/Initializers/experiments/results/cmeans_summary_table.csv
@@ -1,7 +1,0 @@
-,db,db,db,kl_divergence,kl_divergence,kl_divergence,silhouette,silhouette,silhouette,time,time,time
-n_samples,200,500,1000,200,500,1000,200,500,1000,200,500,1000
-method,,,,,,,,,,,,
-cmeans_accurate,0.5008 ± 0.0238,0.5478 ± 0.0251,0.5395 ± 0.0236,0.1827 ± 0.1029,0.2710 ± 0.0046,0.1640 ± 0.0924,0.6325 ± 0.0098,0.5966 ± 0.0131,0.6116 ± 0.0101,0.0021 ± 0.0007,0.0058 ± 0.0012,0.0163 ± 0.0010
-cmeans_fast,0.5008 ± 0.0238,0.5477 ± 0.0250,0.5396 ± 0.0236,0.1828 ± 0.1029,0.2709 ± 0.0046,0.1640 ± 0.0924,0.6325 ± 0.0098,0.5966 ± 0.0131,0.6116 ± 0.0101,0.0013 ± 0.0002,0.0030 ± 0.0012,0.0048 ± 0.0009
-initial,nan,nan,nan,5.8732 ± 5.1260,9.1348 ± 7.7590,8.3692 ± 5.1313,nan,nan,nan,nan,nan,nan
-true,58.0401 ± 106.8217,56.2738 ± 48.2838,167.3528 ± 361.3182,nan,nan,nan,-0.0644 ± 0.0191,-0.0346 ± 0.0069,-0.0388 ± 0.0090,nan,nan,nan
