--- rework_pysatl_mpest/Initializers/experiments/results/combined_500_samples.csv
@@ -1,7 +1,0 @@
-method,silhouette,db,kl_divergence,time
-cmeans_accurate,0.5966 ± 0.0131,0.5478 ± 0.0251,0.2710 ± 0.0046,0.0058 ± 0.0012
-cmeans_fast,0.5966 ± 0.0131,0.5477 ± 0.0250,0.2709 ± 0.0046,0.0030 ± 0.0012
-initial,nan,nan,9.1348 ± 7.7590,nan
-kmeans_accurate,0.6136 ± 0.0111,0.5313 ± 0.0285,0.2267 ± 0.0920,0.0104 ± 0.0015
-kmeans_fast,0.6137 ± 0.0112,0.5315 ± 0.0289,0.2267 ± 0.0922,0.0094 ± 0.0062
-true,-0.0346 ± 0.0069,56.2738 ± 48.2838,nan,nan
