--- rework_pysatl_mpest/Initializers/experiments/results/combined_summary_table.csv
@@ -1,9 +1,0 @@
-,db,db,db,kl_divergence,kl_divergence,kl_divergence,silhouette,silhouette,silhouette,time,time,time
-n_samples,200,500,1000,200,500,1000,200,500,1000,200,500,1000
-method,,,,,,,,,,,,
-cmeans_accurate,0.5008 ± 0.0238,0.5478 ± 0.0251,0.5395 ± 0.0236,0.1827 ± 0.1029,0.2710 ± 0.0046,0.1640 ± 0.0924,0.6325 ± 0.0098,0.5966 ± 0.0131,0.6116 ± 0.0101,0.0021 ± 0.0007,0.0058 ± 0.0012,0.0163 ± 0.0010
-cmeans_fast,0.5008 ± 0.0238,0.5477 ± 0.0250,0.5396 ± 0.0236,0.1828 ± 0.1029,0.2709 ± 0.0046,0.1640 ± 0.0924,0.6325 ± 0.0098,0.5966 ± 0.0131,0.6116 ± 0.0101,0.0013 ± 0.0002,0.0030 ± 0.0012,0.0048 ± 0.0009
-initial,nan,nan,nan,5.8732 ± 5.1260,9.1348 ± 7.7590,8.3692 ± 5.1313,nan,nan,nan,nan,nan,nan
-kmeans_accurate,0.5132 ± 0.0189,0.5313 ± 0.0285,0.5265 ± 0.0183,0.3400 ± 0.0736,0.2267 ± 0.0920,0.1762 ± 0.0886,0.6193 ± 0.0171,0.6136 ± 0.0111,0.6352 ± 0.0070,0.0040 ± 0.0006,0.0104 ± 0.0015,0.0262 ± 0.0012
-kmeans_fast,0.5084 ± 0.0180,0.5315 ± 0.0289,0.5251 ± 0.0193,0.3400 ± 0.0736,0.2267 ± 0.0922,0.1763 ± 0.0886,0.6200 ± 0.0181,0.6137 ± 0.0112,0.6358 ± 0.0073,0.0613 ± 0.1288,0.0094 ± 0.0062,0.0142 ± 0.0014
-true,58.0401 ± 106.8217,56.2738 ± 48.2838,167.3528 ± 361.3182,nan,nan,nan,-0.0644 ± 0.0191,-0.0346 ± 0.0069,-0.0388 ± 0.0090,nan,nan,nan
