--- rework_pysatl_mpest/Initializers/experiments/results/kl_summary.csv
@@ -1,7 +1,0 @@
-method,200,500,1000
-cmeans_accurate,0.1827 ± 0.1029,0.2710 ± 0.0046,0.1640 ± 0.0924
-cmeans_fast,0.1828 ± 0.1029,0.2709 ± 0.0046,0.1640 ± 0.0924
-initial,5.8732 ± 5.1260,9.1348 ± 7.7590,8.3692 ± 5.1313
-kmeans_accurate,0.3400 ± 0.0736,0.2267 ± 0.0920,0.1762 ± 0.0886
-kmeans_fast,0.3400 ± 0.0736,0.2267 ± 0.0922,0.1763 ± 0.0886
-true,nan,nan,nan
