--- rework_pysatl_mpest/Initializers/experiments/results/time_summary.csv
@@ -1,5 +1,0 @@
-method,200,500,1000
-cmeans_accurate,0.0021 ± 0.0007,0.0058 ± 0.0012,0.0163 ± 0.0010
-cmeans_fast,0.0013 ± 0.0002,0.0030 ± 0.0012,0.0048 ± 0.0009
-kmeans_accurate,0.0040 ± 0.0006,0.0104 ± 0.0015,0.0262 ± 0.0012
-kmeans_fast,0.0613 ± 0.1288,0.0094 ± 0.0062,0.0142 ± 0.0014
